services/api/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.36
alembic>=1.13.1

# Validation and Serialization
pydantic>=2.9.0
pydantic-settings>=2.4.0
email-validator>=2.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AWS SDK
boto3==1.34.34

# Machine Learning (pinned to versions with wheels for Python 3.13)
scikit-learn==1.7.0
statsmodels==0.14.4
pandas==2.3.0
numpy==2.3.0

# Computer Vision
ultralytics==8.1.11
opencv-python>=4.9.0.80
Pillow==10.4.0

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
hypothesis==6.98.3

# Utilities
python-dotenv==1.0.1
requests==2.31.0
reportlab==4.2.2
